Hymini STM32F4: Rename files for better conformance to naming conventions
